+++ src/broadcast.jl
@@ -8,3 +8,10 @@ for f in :[%,
   @eval Base.$f(a::Broadcasted, b) = Broadcasted(broadcast_($f, unwrap(a), b))
   @eval Base.$f(b, a::Broadcasted) = Broadcasted(broadcast_($f, b, unwrap(a)))
 end
+
+for f in :[σ, sigmoid, relu, softmax, log_softmax].args
+  # copy from TakingBroadcastSeriously
+  @eval $f(a::Broadcasted...) = Broadcasted(broadcast_($f, unwrap.(a)...))
+  @eval $f(a::Broadcasted, b) = Broadcasted(broadcast_($f, unwrap(a), b))
+  @eval $f(b, a::Broadcasted) = Broadcasted(broadcast_($f, b, unwrap(a)))
+end