+++ src/ndarray.jl
@@ -1075,16 +1075,21 @@ _broadcast_target(sig::Expr) = sig.args[2].args[].args[end]
 Generate docstring from function signature
 """
 function _docsig(fname::Symbol, sig::Expr)
-  s = get(_nddoc, fname, "")
-  !isempty(s) && return s
-
   if fname !== :broadcast_
+    s = get(_nddoc, fname, "")
+    !isempty(s) && return s
     "    $sig"
   else
     name = _broadcast_target(sig)
-    sig_ = Expr(:call, Symbol(name, "."), sig.args[3:end]...)
-    str = "    $sig_"
-    @eval @doc $str $name
+    str = get(_nddoc, name, "")
+    _nddoc[name] = false  # change to false, denote docstring has been set up
+    if isempty(str)
+      sig_ = Expr(:call, Symbol(name, "."), sig.args[3:end]...)
+      str = "    $sig_"
+    end
+    if str ≠ false
+      @eval @doc $str $name
+    end
     ""
   end
 end
@@ -1222,6 +1227,67 @@ julia> mx.expand_dims(x, 2)
 @_remap broadcast_(::typeof(asinh), x::NDArray) arcsinh(x)
 @_remap broadcast_(::typeof(acosh), x::NDArray) arccosh(x)
 @_remap broadcast_(::typeof(atanh), x::NDArray) arctanh(x)
+
+# activation functions
+_nddoc[:σ] = _nddoc[:sigmoid] = doc"""
+    σ.(x::NDArray)
+    sigmoid.(x::NDArray)
+
+Computes sigmoid of x element-wise.
+
+```math
+σ(x) = \frac{1}{(1 + exp(-x))}
+```
+
+The storage type of `sigmoid` output is always dense.
+"""
+@_remap broadcast_(::typeof(σ), x::NDArray)       sigmoid(x)
+@_remap broadcast_(::typeof(sigmoid), x::NDArray) sigmoid(x)
+
+_nddoc[:relu] = doc"""
+    relu.(x::NDArray)
+
+Computes rectified linear.
+
+```math
+\max(x, 0)
+```
+"""
+@_remap broadcast_(::typeof(relu), x::NDArray) relu(x)
+
+_nddoc[:softmax] = doc"""
+    softmax.(x::NDArray, [dim = ndims(x)])
+
+Applies the softmax function.
+
+The resulting array contains elements in the range `(0, 1)`
+and the elements along the given axis sum up to 1.
+
+```math
+softmax(\mathbf{z})_j = \frac{e^{z_j}}{\sum_{k=1}^K e^{z_k}}
+```
+"""
+@_remap broadcast_(::typeof(softmax), x::NDArray) softmax(x; axis = -ndims(x))
+@_remap broadcast_(::typeof(softmax), x::NDArray, dim::Int) softmax(x; axis = -dim)
+
+_nddoc[:log_softmax] = """
+    log_softmax.(x::NDArray, [dim = ndims(x)])
+
+Computes the log softmax of the input.
+This is equivalent to computing softmax followed by log.
+
+julia> x
+2×3 mx.NDArray{Float64,2} @ CPU0:
+ 1.0  2.0  0.1
+ 0.1  2.0  1.0
+
+julia> mx.log_softmax.(x)
+2×3 mx.NDArray{Float64,2} @ CPU0:
+ -1.41703  -0.41703  -2.31703
+ -2.31703  -0.41703  -1.41703
+"""
+@_remap broadcast_(::typeof(log_softmax), x::NDArray) log_softmax(x; axis = -ndims(x))
+@_remap broadcast_(::typeof(log_softmax), x::NDArray, dim::Int) log_softmax(x; axis = -dim)
 
 ################################################################################
 # remapping to solving type unstablility
@@ -1383,6 +1449,12 @@ const _op_import_bl = [  # import black list; do not import these funcs
     "arcsinh",
     "arccosh",
     "arctanh",
+
+    # activation
+    "sigmoid",
+    "relu",
+    "softmax",
+    "log_softmax",
 ]
 
 macro _import_ndarray_functions()